+++ test/IntegrationTest/CommonTestCases.ps1
@@ -83,6 +83,12 @@ $CommonTestCases = @(
     }
     @{ Name = "Contains app context"; TestBlock = {
             param($TestSetup, $TestType, $SentryEvent, $RunResult)
+
+            if ($TestType -eq "crash-capture" -and $TestSetup.Platform -eq "iOS") {
+                Set-ItResult -Skipped -Because "app context is not synced to sentry-cocoa on iOS"
+                return
+            }
+
             $SentryEvent.contexts.app | Should -Not -BeNullOrEmpty
         }
     }