+++ CHANGELOG.md
@@ -12,18 +12,27 @@
 
 ### Dependencies
 
-- Bump Cocoa SDK from v9.4.0 to v9.5.1 ([#2540](https://github.com/getsentry/sentry-unity/pull/2540), [#2558](https://github.com/getsentry/sentry-unity/pull/2558), [#2565](https://github.com/getsentry/sentry-unity/pull/2565))
-  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#951)
-  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.4.0...9.5.1)
+- Bump Cocoa SDK from v9.4.0 to v9.6.0 ([#2540](https://github.com/getsentry/sentry-unity/pull/2540), [#2558](https://github.com/getsentry/sentry-unity/pull/2558), [#2565](https://github.com/getsentry/sentry-unity/pull/2565), [#2575](https://github.com/getsentry/sentry-unity/pull/2575))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#960)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.4.0...9.6.0)
 - Bump Native SDK from v0.12.6 to v0.13.1 ([#2543](https://github.com/getsentry/sentry-unity/pull/2543), [#2550](https://github.com/getsentry/sentry-unity/pull/2550), [#2560](https://github.com/getsentry/sentry-unity/pull/2560), [#2567](https://github.com/getsentry/sentry-unity/pull/2567))
   - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0131)
   - [diff](https://github.com/getsentry/sentry-native/compare/0.12.6...0.13.1)
+<<<<<<< deps/modules/sentry-java/8.34.0
 - Bump Java SDK from v8.32.0 to v8.34.0 ([#2555](https://github.com/getsentry/sentry-unity/pull/2555), [#2574](https://github.com/getsentry/sentry-unity/pull/2574))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8340)
   - [diff](https://github.com/getsentry/sentry-java/compare/8.32.0...8.34.0)
 - Bump CLI from v3.2.0 to v3.2.3 ([#2561](https://github.com/getsentry/sentry-unity/pull/2561), [#2563](https://github.com/getsentry/sentry-unity/pull/2563), [#2568](https://github.com/getsentry/sentry-unity/pull/2568))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#323)
   - [diff](https://github.com/getsentry/sentry-cli/compare/3.2.0...3.2.3)
+=======
+- Bump Java SDK from v8.32.0 to v8.33.0 ([#2555](https://github.com/getsentry/sentry-unity/pull/2555))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8330)
+  - [diff](https://github.com/getsentry/sentry-java/compare/8.32.0...8.33.0)
+- Bump CLI from v3.2.0 to v3.3.0 ([#2561](https://github.com/getsentry/sentry-unity/pull/2561), [#2563](https://github.com/getsentry/sentry-unity/pull/2563), [#2568](https://github.com/getsentry/sentry-unity/pull/2568), [#2576](https://github.com/getsentry/sentry-unity/pull/2576))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#330)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/3.2.0...3.3.0)
+>>>>>>> main
 
 ## 4.1.0